+++ src/app/service/properties/ejector.ts
