+++ src/app/service/utilities/window.service.ts
