+++ src/app/service/valueBindingSystem/pairFormBinding.service.ts
@@ -1,0 +1,9 @@
+import { Injectable } from "@angular/core";
+import { ValueBindingService } from "./valueBinding.service";
+
+@Injectable({providedIn: 'root'})
+export class PairFormBindingService extends ValueBindingService {
+  constructor() {
+    super();
+  }
+}
